tech: move models to database and repository to app
