+++ src/app/repository/auth.ts
@@ -1,4 +1,4 @@
 import { AuthEntity } from '#core/auth/entity'
-import { GenericRepository } from '#shared/repository'
+import { GenericRepository } from '#app/repository/generic'
 
 export type AuthRepository = GenericRepository<AuthEntity>
