+++ src/app/repository/building.ts
@@ -1,4 +1,4 @@
 import { BuildingEntity } from '#core/building/entity'
-import { GenericRepository } from '#shared/repository'
+import { GenericRepository } from '#app/repository/generic'
 
 export type BuildingRepository = GenericRepository<BuildingEntity>
