+++ src/app/repository/city.ts
@@ -1,4 +1,4 @@
 import { CityEntity } from '#core/city/entity'
-import { GenericRepository } from '#shared/repository'
+import { GenericRepository } from '#app/repository/generic'
 
 export type CityRepository = GenericRepository<CityEntity>
