+++ src/app/repository/player.ts
@@ -1,4 +1,4 @@
 import { PlayerEntity } from '#core/player/entity'
-import { GenericRepository } from '#shared/repository'
+import { GenericRepository } from '#app/repository/generic'
 
 export type PlayerRepository = GenericRepository<PlayerEntity>
