+++ src/app/repository/technology.ts
@@ -1,4 +1,4 @@
 import { TechnologyEntity } from '#core/technology/entity'
-import { GenericRepository } from '#shared/repository'
+import { GenericRepository } from '#app/repository/generic'
 
 export type TechnologyRepository = GenericRepository<TechnologyEntity>
