+++ src/database/auth/document.ts
