+++ src/database/building/document.ts
