+++ src/database/repository.ts
@@ -1,26 +1,25 @@
 import { AuthErrors } from '#core/auth/errors'
-import { AuthRepository } from '#core/auth/model'
-import { AuthModel } from '#core/auth/model/document'
-import { MongoAuthRepository } from '#core/auth/model/mongo'
+import { AuthRepository } from '#app/repository/auth'
+import { AuthModel } from '#database/auth/document'
+import { MongoAuthRepository } from '#database/auth/repository'
 import { BuildingErrors } from '#core/building/errors'
-import { BuildingRepository } from '#core/building/model'
-import { BuildingModel } from '#core/building/model/document'
-import { MongoBuildingRepository } from '#core/building/model/mongo'
+import { BuildingRepository } from '#app/repository/building'
+import { BuildingModel } from '#database/building/document'
+import { MongoBuildingRepository } from '#database/building/repository'
 import { CityErrors } from '#core/city/errors'
-import { CityRepository } from '#core/city/model'
-import { CityModel } from '#core/city/model/document'
-import { MongoCityRepository } from '#core/city/model/mongo'
+import { CityRepository } from '#app/repository/city'
+import { CityModel } from '#database/city/document'
+import { MongoCityRepository } from '#database/city/repository'
 import { PlayerErrors } from '#core/player/errors'
-import { PlayerRepository } from '#core/player/model'
-import { PlayerModel } from '#core/player/model/document'
-import { MongoPlayerRepository } from '#core/player/model/mongo'
+import { PlayerRepository } from '#app/repository/player'
+import { PlayerModel } from '#database/player/document'
+import { MongoPlayerRepository } from '#database/player/repository'
 import { TechnologyErrors } from '#core/technology/errors'
-import { TechnologyRepository } from '#core/technology/model'
-import { TechnologyModel } from '#core/technology/model/document'
-import { MongoTechnologyRepository } from '#core/technology/model/mongo'
-import { Repository } from '#shared/repository'
+import { TechnologyRepository } from '#app/repository/technology'
+import { TechnologyModel } from '#database/technology/document'
+import { MongoTechnologyRepository } from '#database/technology/repository'
+import { Repository } from '#app/repository/generic'
 import { mongoose } from '@typegoose/typegoose'
-
 
 export class MongoRepository implements Repository {
   auth: AuthRepository
