+++ src/web/middleware/auth.ts
@@ -1,5 +1,5 @@
-import { AppQueries } from '#app/query'
-import { Factory } from '#core/factory'
+import { Queries } from '#app/queries'
+import { Factory } from '#app/factory'
 import {
   NextFunction, Request, Response
 } from 'express'
@@ -13,7 +13,7 @@ export const authMiddleware = async (req: Request<unknown>, res: Response<unknow
     })
   }
   try {
-    const queries = new AppQueries({ repository: Factory.getRepository() })
+    const queries = new Queries({ repository: Factory.getRepository() })
     const { player_id } = await queries.authorize({ token })
     res.locals.player_id = player_id
     next()
